removed property tests to single tests
